Move some more stuff around
